[geogram] Update to 1.7.4

+++ src/lib/geogram/image/image_library.cpp
@@ -244,8 +244,8 @@ namespace GEO {
             char *pData = (char *) ::GlobalLock((HGLOBAL)handle);
             BITMAPINFOHEADER header ;
             header.biSize          = sizeof(BITMAPINFOHEADER);
-            header.biWidth         = image->width() ;
-            header.biHeight        = image->height() ;
+            header.biWidth         = (LONG)(image->width()) ;
+	    header.biHeight        = (LONG)(image->height()) ;
             header.biPlanes        = 1 ;
             header.biBitCount      = 24 ;
             header.biCompression   = BI_RGB ;
